requirements.txt
python-telegram-bot==20.3
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0
pytest==7.4.3
loguru==0.7.2